requirements.txt
Adafruit-Blinka==6.15.0
adafruit-circuitpython-busdevice==5.1.0
adafruit-circuitpython-rgb-display==3.10.8
Adafruit-PureIO==1.1.9
font-roboto==0.0.1
fonts==0.0.3
numpy==1.16.2
Pillow==8.4.0
RPi.GPIO==0.7.0
ST7789==0.0.4
spidev==3.5
